feat(#519): ICA: format-on-save — Thin adapter module fragmentation
