+++ .pi/extensions/format-on-save/eslint.mts
@@ -1,4 +1,12 @@
-import type { EslintDiagnostic } from "./types.mts";
+/** ESLint diagnostic entry parsed from JSON output. */
+export interface EslintDiagnostic {
+	file: string;
+	line: number;
+	column: number;
+	severity: "Error" | "Warning";
+	message: string;
+	ruleId: string | null;
+}
 
 /**
  * Parse ESLint JSON output into diagnostics array.
